Keep permalink change through rewrite rule flushes
The filtering of the rewrite rules needs to happen on `init`, so that any call from WP core, another plugin or theme, or a user visiting Settings -> Permalinks flushes the rewrite rules with our change included.

Since the `init` hook is now used, the deactivation hook needs improving too.

Fixes #1